requirements.txt
awscli==1.16.201
awsebcli==3.15.2
blessed==1.15.0
botocore==1.12.191
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
idna==2.7
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
pathspec==0.5.9
Pillow==6.0.0
pyasn1==0.4.5
Pygments==2.3.1
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.2
wcwidth==0.1.7
websocket-client==0.56.0
